+++ src/Shared/HttpClientHelpers.cs
@@ -11,7 +11,7 @@ namespace System.Net.Http;
 internal static class HttpClientHelpers
 {
     // See https://github.com/open-telemetry/opentelemetry-proto/pull/781
-    private const int DefaultMessageSizeLimit = 4 * 1024 * 1024; // 4MiB
+    internal const int DefaultMessageSizeLimit = 4 * 1024 * 1024; // 4MiB
 
     internal static string? TryGetResponseBodyAsString(HttpResponseMessage? httpResponse, CancellationToken cancellationToken)
         => GetResponseBodyAsString(allowTruncation: true, DefaultMessageSizeLimit, httpResponse, cancellationToken);